% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{gc_get_colour}
\alias{gc_get_colour}
\title{Fetching GlobColour products}
\usage{
gc_get_colour(query, openFile = FALSE)
}
\arguments{
\item{query}{character or vector giving the file directories.}

\item{openFile}{logical, if TRUE, function opens previously downloaded file (see Details). If FALSE, files are saved to working directory (default).}
}
\description{
Function downloads the requested GlobColour product, given as file directories created by \code{\link{gc_get_option}}.
}
\details{
Required file directories can either be generated by \code{\link{gc_get_colour}} or, if known, manually created. Note
that the quantity of file directories is dramatically correlated with download time.

If \code{openFile} is set TRUE the function creates a temporary file (see \code{\link[base]{tempfile}}), which is immediately opend after 
download (see Examples). Method requires \code{length(query) < 2}. For further details see \code{\link[ncdf4]{nc_open}} and \code{\link[png]{readPNG}}.
}
\examples{
\dontrun{

#Provide user information
gc_set_user("ftp_gc_DDummy", "DDummy_1234")

#Create directory for quick-look-file.png
qu <- gc_get_option(zone = "GLOB",
                   sens = "merged",
                   binp = "day",
                   date = "2017/08/19",
                   type = "KD490",
                   fext = "png")
             
#Download and assign file to object                          
test <- gc_get_colour(qu, openFile = TRUE)

#Plotting
grid::grid.raster(test)
}
}
