% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{gc_get_option}
\alias{gc_get_option}
\title{Get vector of available GlobColour products}
\usage{
gc_get_option(zone, sens, binp, date, type, fext, prle, reso)
}
\arguments{
\item{zone}{character indicating the requested region (mendatory). Use "GLOB" for global coverage or "EURO" for the european area (see Details).}

\item{sens}{character indicating a satellite acronym (mendatory). Accepted values are "meris", "viirsn" ,"olcia", "modis", "seawifs", or
"merged".}

\item{binp}{character indicating the binning period (mendatory). Accepted values are "month", "8-day", or "day".}

\item{date}{character defining the requested date (mendatory). Required format is "yyyy/mm/dd". Use "latest" for 
most current data (see Details).}

\item{type}{character or vector of product types (optional). See Details for provided GlobColour products.}

\item{fext}{character or vector of file extentions (optional). Provided data formats are "nc" (NetCDF4) and "png" (PortableNetworkGraphic).}

\item{prle}{character or vector indicating the product level (optional). Accepted values are "L3m" (mapped grid) and "L3b" (ISIN grid).}

\item{reso}{character or vector specifying the resolution (optional) to "1", "4", "25", and/or "100" km (see Details).}
}
\description{
Function examines available GlobColour products for a certain query and provides the respective file 
paths for direct FTP access via \code{\link{gc_get_colour}}.
}
\details{
The first four arguments (\code{zone}, \code{sens}, \code{binp}, \code{date}) are mendatory, without which the function returns 
a table of accepted values to type in. The other arguments (\code{type}, \code{fext}, \code{prle}, \code{reso}) provide further filtering. 
For additional information see the \href{http://www.globcolour.info/CDR_Docs/GlobCOLOUR_PUG.pdf}{GlobColour Product User's Guide}.

The resolutions provided for the global coverage are 4, 25, and 100 km. The resolution for the europe-specific dataset 
is 1 km only. Values can optionally be set by \code{reso}.

If the requested \code{date} is after the most current date availabe the function automatically updates this attribute. The user
can also type "latest" for automatically fetching the latest release.

The function shows all available types of data when using the mendatory arguments only. The provided data types (e.g. "KD490", "PAR", "CHL1") 
can then be used in \code{type} to accelerate the search process, and much more importantly, to shorten the list of files for download.
}
\examples{
\dontrun{

#Provide user information
gc_set_user("ftp_gc_DDummy", "DDummy_1234")

#Accepted values to use for arguments (mendatory vs. optional)
gc_get_option()

#Latest available ocean optics (merged) on global coverage and daily intervals
gc_get_option(zone = "GLOB",
              sens = "merged",
              binp = "day",
              date = "latest")

#Further filtering options
gc_get_option(zone = "GLOB",
              sens = "merged",
              binp = "day",
              date = "latest",                          
              type = c("KD490", "PAR"),
              fext = c("nc"),
              reso = c(4, 25))
}
}
