% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{gc_set_user}
\alias{gc_set_user}
\title{Set username and password used for authentication by ACRI}
\usage{
gc_set_user(username, password)
}
\arguments{
\item{username}{character string, username registered at \href{http://hermes.acri.fr/index.php?class=ftp_access}{hermes.acri.fr}.}

\item{password}{character string, password provided by ACRI after registration.}
}
\description{
Function to set username and password used for authentication by ACRI.
}
\details{
Function sets environment variables (ACRI_FTP_USERNAME, ACRI_FTP_PASSWORD) used for FTP connection to the GlobColour database.
Previous registration for access on the GlobColour archive is required.
}
\examples{
\dontrun{

gc_set_user("ftp_gc_DDummy", "DDummy_1234")
Sys.getenv(c("ACRI_FTP_USERNAME", "ACRI_FTP_PASSWORD"))
}
}
